refactor: DiagonalMatrixのビルダの内容を,別クラスに分離する
